refactor: extract ExcelHelper and TaxSummarySheetBuilder from Output
Move tax summary sheet logic into TaxSummarySheetBuilder and reusable
Excel utilities (column letter conversion, header scanning) into
ExcelHelper. Keeps Output.cs focused on generic output dispatch.
