+++ StatementParser/TaxReporterCLI/Output.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
@@ -41,7 +41,8 @@ namespace TaxReporterCLI
                 sheetMeta[@group.Key] = (sheet, @group.Value.Count);
             }
 
-            wb1.Add(CreateTaxSummarySheet(wb1, sheetMeta));
+            var summaryBuilder = new TaxSummarySheetBuilder();
+            wb1.Add(summaryBuilder.Build(wb1, sheetMeta));
 
             wb1.Write(file);
         }
@@ -103,144 +104,6 @@ namespace TaxReporterCLI
 			}
 
 			return row;
-		}
-
-		private ISheet CreateTaxSummarySheet(XSSFWorkbook workbook, Dictionary<string, (ISheet Sheet, int DataRowCount)> sheetMeta)
-		{
-			var summarySheet = workbook.CreateSheet("Tax Summary");
-
-			// CZK currency format
-			var czkFormat = workbook.CreateDataFormat().GetFormat("#,##0.00 \"Kč\"");
-			var czkStyle = workbook.CreateCellStyle();
-			czkStyle.DataFormat = czkFormat;
-
-			// Row 0: Header
-			var headerRow = summarySheet.CreateRow(0);
-			headerRow.CreateCell(0).SetCellValue("Section");
-			headerRow.CreateCell(1).SetCellValue("Daily Rate (CZK)");
-			headerRow.CreateCell(2).SetCellValue("Yearly Rate (CZK)");
-
-			// Define sections: (label, sheetName, dailyPrefix, dailySuffix, yearlyPrefix, yearlySuffix)
-			var sections = new[]
-			{
-				("Dividends - Income", "DividendTransactionView", "Income", "per Day", "Income", "per Year"),
-				("Sales - Profit", "SaleTransactionView", "Profit", "per Day", "Profit", "per Year"),
-				("ESPP - Profit", "ESPPTransactionView", "Total Profit", "per Day", "Total Profit", "per Year"),
-				("Deposits - Price", "DepositTransactionView", "Total Price", "per Day", "Total Price", "per Year"),
-			};
-
-			int summaryRowIndex = 1;
-			foreach (var (label, sheetName, dailyPrefix, dailySuffix, yearlyPrefix, yearlySuffix) in sections)
-			{
-				var row = summarySheet.CreateRow(summaryRowIndex);
-				row.CreateCell(0).SetCellValue(label);
-
-				if (sheetMeta.TryGetValue(sheetName, out var meta) && meta.DataRowCount > 0)
-				{
-					WriteSumFormulaCell(row, 1, meta.Sheet, meta.DataRowCount, dailyPrefix, dailySuffix, czkStyle);
-					WriteSumFormulaCell(row, 2, meta.Sheet, meta.DataRowCount, yearlyPrefix, yearlySuffix, czkStyle);
-				}
-				else
-				{
-					var dailyCell = row.CreateCell(1);
-					dailyCell.SetCellValue(0);
-					dailyCell.CellStyle = czkStyle;
-					var yearlyCell = row.CreateCell(2);
-					yearlyCell.SetCellValue(0);
-					yearlyCell.CellStyle = czkStyle;
-				}
-
-				summaryRowIndex++;
-			}
-
-			// Total row (SUM of the 4 section rows above)
-			var totalRow = summarySheet.CreateRow(summaryRowIndex);
-			totalRow.CreateCell(0).SetCellValue("Total");
-			var totalDailyCell = totalRow.CreateCell(1);
-			totalDailyCell.SetCellFormula("SUM(B2:B5)");
-			totalDailyCell.CellStyle = czkStyle;
-			var totalYearlyCell = totalRow.CreateCell(2);
-			totalYearlyCell.SetCellFormula("SUM(C2:C5)");
-			totalYearlyCell.CellStyle = czkStyle;
-			summaryRowIndex++;
-
-			// Empty row
-			summaryRowIndex++;
-
-			// Dividends - Tax withheld (separate from total)
-			var taxRow = summarySheet.CreateRow(summaryRowIndex);
-			taxRow.CreateCell(0).SetCellValue("Dividends - Tax withheld");
-
-			if (sheetMeta.TryGetValue("DividendTransactionView", out var divMeta) && divMeta.DataRowCount > 0)
-			{
-				WriteSumFormulaCell(taxRow, 1, divMeta.Sheet, divMeta.DataRowCount, "Tax", "per Day", czkStyle);
-				WriteSumFormulaCell(taxRow, 2, divMeta.Sheet, divMeta.DataRowCount, "Tax", "per Year", czkStyle);
-			}
-			else
-			{
-				var dailyCell = taxRow.CreateCell(1);
-				dailyCell.SetCellValue(0);
-				dailyCell.CellStyle = czkStyle;
-				var yearlyCell = taxRow.CreateCell(2);
-				yearlyCell.SetCellValue(0);
-				yearlyCell.CellStyle = czkStyle;
-			}
-
-			// Auto-size columns
-			for (int col = 0; col <= 2; col++)
-			{
-				summarySheet.AutoSizeColumn(col);
-			}
-
-			return summarySheet;
-		}
-
-		private void WriteSumFormulaCell(IRow row, int cellIndex, ISheet sourceSheet, int dataRowCount, string headerPrefix, string headerSuffix, ICellStyle style)
-		{
-			var colIndex = FindColumnIndex(sourceSheet, headerPrefix, headerSuffix);
-			var cell = row.CreateCell(cellIndex);
-			cell.CellStyle = style;
-			if (colIndex >= 0)
-			{
-				var colLetter = GetExcelColumnLetter(colIndex);
-				var formula = $"SUM({sourceSheet.SheetName}!{colLetter}2:{colLetter}{dataRowCount + 1})";
-				cell.SetCellFormula(formula);
-			}
-			else
-			{
-				cell.SetCellValue(0);
-			}
-		}
-
-		private static int FindColumnIndex(ISheet sheet, string startsWith, string endsWith)
-		{
-			var headerRow = sheet.GetRow(0);
-			if (headerRow == null) return -1;
-
-			for (int i = 0; i < headerRow.LastCellNum; i++)
-			{
-				var cell = headerRow.GetCell(i);
-				if (cell == null) continue;
-
-				var value = cell.StringCellValue;
-				if (value != null && value.StartsWith(startsWith) && value.EndsWith(endsWith))
-				{
-					return i;
-				}
-			}
-
-			return -1;
-		}
-
-		private static string GetExcelColumnLetter(int columnIndex)
-		{
-			var result = "";
-			while (columnIndex >= 0)
-			{
-				result = (char)('A' + columnIndex % 26) + result;
-				columnIndex = columnIndex / 26 - 1;
-			}
-			return result;
 		}
 
 		private IDictionary<PropertyInfo, string> CollectPublicProperties(Object instance)
